remove old code & change to new api
