+++ src/Console/Command/SubscriptionPauseCommand.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Patchlevel\EventSourcing\Console\Command;
 
+use Patchlevel\EventSourcing\Subscription\Engine\Command\Pause;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -17,7 +18,10 @@ final class SubscriptionPauseCommand extends SubscriptionCommand
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $criteria = $this->subscriptionEngineCriteria($input);
-        $this->engine->pause($criteria);
+        $this->engine->run(new Pause(
+            $criteria->ids,
+            $criteria->groups,
+        ));
 
         return 0;
     }
