+++ src/Console/Command/SubscriptionRemoveCommand.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Patchlevel\EventSourcing\Console\Command;
 
 use Patchlevel\EventSourcing\Console\OutputStyle;
+use Patchlevel\EventSourcing\Subscription\Engine\Command\Remove;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -27,7 +28,7 @@ final class SubscriptionRemoveCommand extends SubscriptionCommand
             }
         }
 
-        $this->engine->remove($criteria);
+        $this->engine->run(new Remove($criteria->ids, $criteria->groups));
 
         return 0;
     }
