--- src/Subscription/Engine/CanRefreshSubscriptions.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Patchlevel\EventSourcing\Subscription\Engine;
-
-interface CanRefreshSubscriptions
-{
-    public function refresh(SubscriptionEngineCriteria|null $criteria = null): Result;
-}
