+++ src/Subscription/Engine/DefaultSubscriptionEngine.php
@@ -4,38 +4,55 @@ declare(strict_types=1);
 
 namespace Patchlevel\EventSourcing\Subscription\Engine;
 
-use Patchlevel\EventSourcing\Message\Message;
+use InvalidArgumentException;
 use Patchlevel\EventSourcing\Subscription\Cleanup\Cleaner;
+use Patchlevel\EventSourcing\Subscription\Engine\Command\Boot;
+use Patchlevel\EventSourcing\Subscription\Engine\Command\Command;
+use Patchlevel\EventSourcing\Subscription\Engine\Command\Pause;
+use Patchlevel\EventSourcing\Subscription\Engine\Command\Reactivate;
+use Patchlevel\EventSourcing\Subscription\Engine\Command\Refresh;
+use Patchlevel\EventSourcing\Subscription\Engine\Command\Remove;
+use Patchlevel\EventSourcing\Subscription\Engine\Command\Run;
+use Patchlevel\EventSourcing\Subscription\Engine\Command\Setup;
+use Patchlevel\EventSourcing\Subscription\Engine\Command\Teardown;
+use Patchlevel\EventSourcing\Subscription\Engine\Event\OnCommand;
+use Patchlevel\EventSourcing\Subscription\Engine\Event\OnResult;
+use Patchlevel\EventSourcing\Subscription\Engine\Event\OnSubscriptions;
+use Patchlevel\EventSourcing\Subscription\Engine\Handler\BootHandler;
+use Patchlevel\EventSourcing\Subscription\Engine\Handler\Handler;
+use Patchlevel\EventSourcing\Subscription\Engine\Handler\PauseHandler;
+use Patchlevel\EventSourcing\Subscription\Engine\Handler\ReactivateHandler;
+use Patchlevel\EventSourcing\Subscription\Engine\Handler\RefreshHandler;
+use Patchlevel\EventSourcing\Subscription\Engine\Handler\RemoveHandler;
+use Patchlevel\EventSourcing\Subscription\Engine\Handler\RunHandler;
+use Patchlevel\EventSourcing\Subscription\Engine\Handler\SetupHandler;
+use Patchlevel\EventSourcing\Subscription\Engine\Handler\TeardownHandler;
+use Patchlevel\EventSourcing\Subscription\Engine\Listener\BatchSubscriber;
+use Patchlevel\EventSourcing\Subscription\Engine\Listener\DetachListener;
+use Patchlevel\EventSourcing\Subscription\Engine\Listener\DiscoverSubscriber;
+use Patchlevel\EventSourcing\Subscription\Engine\Listener\FailSubscriber;
+use Patchlevel\EventSourcing\Subscription\Engine\Listener\RetrySubscriber;
 use Patchlevel\EventSourcing\Subscription\RetryStrategy\ClockBasedRetryStrategy;
-use Patchlevel\EventSourcing\Subscription\RetryStrategy\ConditionalRetryStrategy;
 use Patchlevel\EventSourcing\Subscription\RetryStrategy\NoRetryStrategy;
-use Patchlevel\EventSourcing\Subscription\RetryStrategy\RetryStrategy;
 use Patchlevel\EventSourcing\Subscription\RetryStrategy\RetryStrategyRepository;
-use Patchlevel\EventSourcing\Subscription\RunMode;
-use Patchlevel\EventSourcing\Subscription\Status;
 use Patchlevel\EventSourcing\Subscription\Store\SubscriptionCriteria;
 use Patchlevel\EventSourcing\Subscription\Store\SubscriptionStore;
-use Patchlevel\EventSourcing\Subscription\Subscriber\BatchableSubscriber;
-use Patchlevel\EventSourcing\Subscription\Subscriber\MetadataSubscriberAccessor;
 use Patchlevel\EventSourcing\Subscription\Subscriber\SubscriberAccessorRepository;
 use Patchlevel\EventSourcing\Subscription\Subscription;
 use Psr\Log\LoggerInterface;
-use Throwable;
+use Symfony\Component\EventDispatcher\EventDispatcher;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
-use function array_values;
-use function count;
-use function sprintf;
-
-final class DefaultSubscriptionEngine implements SubscriptionEngine, CanRefreshSubscriptions
+final class DefaultSubscriptionEngine implements SubscriptionEngine
 {
     private SubscriptionManager $subscriptionManager;
 
     private bool $processing = false;
 
-    /** @var array<string, BatchableSubscriber> */
-    private array $batching = [];
-
     private readonly RetryStrategyRepository $retryStrategyRepository;
+
+    /** @var array<class-string<Command>, Handler> */
+    private readonly array $handlers;
 
     public function __construct(
         private readonly MessageLoader $messageLoader,
@@ -44,6 +61,7 @@ final class DefaultSubscriptionEngine implements SubscriptionEngine, CanRefreshS
         RetryStrategyRepository|null $retryStrategyRepository = null,
         private readonly LoggerInterface|null $logger = null,
         private readonly Cleaner|null $cleaner = null,
+        private readonly EventDispatcherInterface $eventDispatcher = new EventDispatcher(),
     ) {
         $this->subscriptionManager = new SubscriptionManager($subscriptionStore);
 
@@ -55,736 +73,157 @@ final class DefaultSubscriptionEngine implements SubscriptionEngine, CanRefreshS
                 'no_retry' => new NoRetryStrategy(),
             ]);
         }
+
+        $cleanupRunner = new CleanupRunner(
+            $this->subscriptionManager,
+            $this->cleaner,
+            $this->logger,
+        );
+
+        $messageProcessor = new MessageProcessor(
+            $this->subscriberRepository,
+            $this->eventDispatcher,
+            $this->logger,
+        );
+
+        $this->handlers = [
+            Boot::class => new BootHandler(
+                $this->messageLoader,
+                $this->subscriptionManager,
+                $this->subscriberRepository,
+                $messageProcessor,
+                $this->eventDispatcher,
+                $this->logger,
+            ),
+            Pause::class => new PauseHandler(
+                $this->subscriptionManager,
+                $this->logger,
+            ),
+            Reactivate::class => new ReactivateHandler(
+                $this->subscriptionManager,
+                $this->subscriberRepository,
+                $this->logger,
+            ),
+            Refresh::class => new RefreshHandler(
+                $this->subscriptionManager,
+                $this->subscriberRepository,
+                $this->logger,
+            ),
+            Remove::class => new RemoveHandler(
+                $this->subscriptionManager,
+                $this->subscriberRepository,
+                $cleanupRunner,
+                $this->logger,
+            ),
+            Run::class => new RunHandler(
+                $this->messageLoader,
+                $this->subscriptionManager,
+                $messageProcessor,
+                $this->eventDispatcher,
+                $this->logger,
+            ),
+            Setup::class => new SetupHandler(
+                $this->messageLoader,
+                $this->subscriptionManager,
+                $this->subscriberRepository,
+                $this->retryStrategyRepository,
+                $this->logger,
+            ),
+            Teardown::class => new TeardownHandler(
+                $this->subscriptionManager,
+                $this->subscriberRepository,
+                $cleanupRunner,
+                $this->logger,
+            ),
+        ];
+
+        $this->eventDispatcher->addSubscriber(
+            new DiscoverSubscriber(
+                $this->messageLoader,
+                $this->subscriptionManager,
+                $this->subscriberRepository,
+                $this->logger,
+            ),
+        );
+
+        $this->eventDispatcher->addSubscriber(
+            new RetrySubscriber(
+                $this->subscriptionManager,
+                $this->subscriberRepository,
+                $this->retryStrategyRepository,
+                $this->logger,
+            ),
+        );
+
+        $this->eventDispatcher->addSubscriber(
+            new BatchSubscriber(
+                $this->subscriberRepository,
+                $this->logger,
+            ),
+        );
+
+        $this->eventDispatcher->addSubscriber(
+            new FailSubscriber(
+                $this->subscriptionManager,
+                $this->subscriberRepository,
+                $this->logger,
+            ),
+        );
+
+        $this->eventDispatcher->addListener(
+            OnCommand::class,
+            new DetachListener(
+                $this->subscriptionManager,
+                $this->subscriberRepository,
+                $this->logger,
+            ),
+            32,
+        );
     }
 
-    public function setup(SubscriptionEngineCriteria|null $criteria = null, bool $skipBooting = false): Result
+    public function run(Command $command): Result
     {
-        $criteria ??= new SubscriptionEngineCriteria();
-
         $this->logger?->info(
-            'Subscription Engine: Start to setup.',
+            'Subscription Engine: ' . $command::class . ' command received.',
         );
 
-        $this->discoverNewSubscriptions();
-        $this->retrySubscriptions($criteria, Status::New);
-
-        return $this->subscriptionManager->findForUpdate(
-            new SubscriptionCriteria(
-                ids: $criteria->ids,
-                groups: $criteria->groups,
-                status: [Status::New],
-            ),
-            function (SubscriptionCollection $subscriptions) use ($skipBooting): Result {
-                if (count($subscriptions) === 0) {
-                    $this->logger?->info('Subscription Engine: No subscriptions to setup, finish setup.');
-
-                    return new Result();
-                }
-
-                /** @var list<Error> $errors */
-                $errors = [];
-
-                $latestIndex = $this->messageLoader->lastIndex();
-
-                foreach ($subscriptions as $subscription) {
-                    $subscriber = $this->subscriber($subscription->id());
-
-                    if (!$subscriber) {
-                        throw SubscriberNotFound::forSubscriptionId($subscription->id());
-                    }
-
-                    $setupMethod = $subscriber->setupMethod();
-
-                    if (!$setupMethod) {
-                        if ($subscription->runMode() === RunMode::FromNow) {
-                            $subscription->changePosition($latestIndex);
-                            $subscription->active();
-                        } else {
-                            $skipBooting ? $subscription->active() : $subscription->booting();
-                        }
-
-                        $this->subscriptionManager->update($subscription);
-
-                        $this->logger?->debug(sprintf(
-                            'Subscription Engine: Subscriber "%s" for "%s" has no setup method, set to %s.',
-                            $subscriber::class,
-                            $subscription->id(),
-                            $subscription->runMode() === RunMode::FromNow || $skipBooting ? 'active' : 'booting',
-                        ));
-
-                        continue;
-                    }
-
-                    try {
-                        $setupMethod();
-
-                        if ($subscription->runMode() === RunMode::FromNow) {
-                            $subscription->changePosition($latestIndex);
-                            $subscription->active();
-                        } else {
-                            $skipBooting ? $subscription->active() : $subscription->booting();
-                        }
-
-                        $this->subscriptionManager->update($subscription);
-
-                        $this->logger?->debug(sprintf(
-                            'Subscription Engine: For Subscriber "%s" for "%s" the setup method has been executed, set to %s.',
-                            $subscriber::class,
-                            $subscription->id(),
-                            $subscription->runMode() === RunMode::FromNow || $skipBooting ? 'active' : 'booting',
-                        ));
-                    } catch (Throwable $e) {
-                        $this->logger?->error(sprintf(
-                            'Subscription Engine: Subscriber "%s" for "%s" has an error in the setup method: %s',
-                            $subscriber::class,
-                            $subscription->id(),
-                            $e->getMessage(),
-                        ));
-
-                        $this->handleError($subscription, $e);
-
-                        $errors[] = new Error(
-                            $subscription->id(),
-                            $e->getMessage(),
-                            $e,
-                        );
-                    }
-                }
-
-                return new Result($errors);
-            },
-        );
-    }
-
-    public function boot(
-        SubscriptionEngineCriteria|null $criteria = null,
-        int|null $limit = null,
-    ): ProcessedResult {
         if ($this->processing) {
+            $this->logger?->error(
+                'Subscription Engine: Already processing, skip.',
+            );
+
             throw new AlreadyProcessing();
         }
 
         $this->processing = true;
-        $this->batching = [];
 
         try {
-            $criteria ??= new SubscriptionEngineCriteria();
+            $handler = $this->handlers[$command::class] ?? null;
 
+            if ($handler === null) {
+                throw new InvalidArgumentException('No handler found for command: ' . $command::class);
+            }
+
+            $this->logger?->debug(
+                'Subscription Engine: ' . $command::class . ' command handled by ' . $handler::class,
+            );
+
+            $event = new OnCommand($command);
+            $this->eventDispatcher->dispatch($event);
+
+            $result = $handler($command);
+
+            $event = new OnResult($command, $result);
+            $this->eventDispatcher->dispatch($event);
+
+            return $result;
+        } finally {
             $this->logger?->info(
-                'Subscription Engine: Start booting.',
+                'Subscription Engine: ' . $command::class . ' command processed.',
             );
 
-            $this->discoverNewSubscriptions();
-            $this->retrySubscriptions($criteria, Status::Booting);
-
-            return $this->subscriptionManager->findForUpdate(
-                new SubscriptionCriteria(
-                    ids: $criteria->ids,
-                    groups: $criteria->groups,
-                    status: [Status::Booting],
-                ),
-                function (SubscriptionCollection $subscriptions) use ($limit): ProcessedResult {
-                    if (count($subscriptions) === 0) {
-                        $this->logger?->info('Subscription Engine: No subscriptions in booting status, finish booting.');
-
-                        return new ProcessedResult(0, true);
-                    }
-
-                    foreach ($subscriptions as $subscription) {
-                        $subscriber = $this->subscriber($subscription->id());
-
-                        if ($subscriber) {
-                            continue;
-                        }
-
-                        $this->logger?->debug(
-                            sprintf(
-                                'Subscription Engine: Subscriber for "%s" not found, skipped.',
-                                $subscription->id(),
-                            ),
-                        );
-
-                        $subscriptions->remove($subscription);
-                    }
-
-                    $startIndex = $subscriptions->lowestPosition();
-
-                    $this->logger?->debug(
-                        sprintf(
-                            'Subscription Engine: Event stream is processed for booting from position %s.',
-                            $startIndex,
-                        ),
-                    );
-
-                    /** @var list<Error> $errors */
-                    $errors = [];
-                    $stream = null;
-                    $messageCounter = 0;
-                    $lastIndex = null;
-
-                    try {
-                        $stream = $this->messageLoader->load($startIndex, $subscriptions->toArray());
-
-                        foreach ($stream as $index => $message) {
-                            $messageCounter++;
-                            $lastIndex = $index;
-
-                            foreach ($subscriptions as $subscription) {
-                                if ($subscription->position() >= $index) {
-                                    $this->logger?->debug(
-                                        sprintf(
-                                            'Subscription Engine: Subscription "%s" is farther than the current position (%d > %d), continue booting.',
-                                            $subscription->id(),
-                                            $subscription->position(),
-                                            $index,
-                                        ),
-                                    );
-
-                                    continue;
-                                }
-
-                                $error = $this->handleMessage($index, $message, $subscription);
-
-                                if (!$error) {
-                                    continue;
-                                }
-
-                                $errors[] = $error;
-
-                                $subscriptions->remove($subscription);
-
-                                if (count($subscriptions) === 0) {
-                                    $this->logger?->info(
-                                        'Subscription Engine: No subscriptions in booting status, finish booting.',
-                                    );
-
-                                    break 2;
-                                }
-                            }
-
-                            $this->logger?->debug(
-                                sprintf(
-                                    'Subscription Engine: Current event stream position for booting: %s',
-                                    $index,
-                                ),
-                            );
-
-                            if ($limit !== null && $messageCounter >= $limit) {
-                                $this->logger?->info(
-                                    sprintf(
-                                        'Subscription Engine: Message limit (%d) reached, finish booting.',
-                                        $limit,
-                                    ),
-                                );
-
-                                return new ProcessedResult(
-                                    $messageCounter,
-                                    false,
-                                    $errors,
-                                );
-                            }
-                        }
-                    } finally {
-                        $stream?->close();
-
-                        if ($lastIndex !== null && $messageCounter > 0) {
-                            foreach ($subscriptions as $subscription) {
-                                $error = $this->ensureCommitBatch($subscription, $lastIndex);
-
-                                if ($error) {
-                                    $errors[] = $error;
-
-                                    $subscriptions->remove($subscription);
-                                }
-
-                                $this->subscriptionManager->update($subscription);
-                            }
-                        }
-                    }
-
-                    $this->logger?->debug('Subscription Engine: End of stream for booting has been reached.');
-
-                    foreach ($subscriptions as $subscription) {
-                        if ($subscription->runMode() === RunMode::Once) {
-                            $subscription->finished();
-                            $this->subscriptionManager->update($subscription);
-
-                            $this->logger?->info(sprintf(
-                                'Subscription Engine: Subscription "%s" run only once and has been set to finished.',
-                                $subscription->id(),
-                            ));
-
-                            continue;
-                        }
-
-                        $subscription->active();
-                        $this->subscriptionManager->update($subscription);
-
-                        $this->logger?->info(sprintf(
-                            'Subscription Engine: Subscription "%s" has been set to active after booting.',
-                            $subscription->id(),
-                        ));
-                    }
-
-                    $this->logger?->info('Subscription Engine: Finish booting.');
-
-                    return new ProcessedResult(
-                        $messageCounter,
-                        true,
-                        $errors,
-                    );
-                },
-            );
-        } finally {
             $this->processing = false;
         }
-    }
-
-    public function run(
-        SubscriptionEngineCriteria|null $criteria = null,
-        int|null $limit = null,
-    ): ProcessedResult {
-        if ($this->processing) {
-            throw new AlreadyProcessing();
-        }
-
-        $this->processing = true;
-        $this->batching = [];
-
-        try {
-            $criteria ??= new SubscriptionEngineCriteria();
-
-            $this->logger?->info('Subscription Engine: Start processing.');
-
-            $this->discoverNewSubscriptions();
-            $this->markDetachedSubscriptions($criteria);
-            $this->retrySubscriptions($criteria, Status::Active);
-
-            return $this->subscriptionManager->findForUpdate(
-                new SubscriptionCriteria(
-                    ids: $criteria->ids,
-                    groups: $criteria->groups,
-                    status: [Status::Active],
-                ),
-                function (SubscriptionCollection $subscriptions) use ($limit): ProcessedResult {
-                    if (count($subscriptions) === 0) {
-                        $this->logger?->info('Subscription Engine: No subscriptions to process, finish processing.');
-
-                        return new ProcessedResult(0, true);
-                    }
-
-                    $startIndex = $subscriptions->lowestPosition();
-
-                    $this->logger?->debug(
-                        sprintf(
-                            'Subscription Engine: Event stream is processed from position %d.',
-                            $startIndex,
-                        ),
-                    );
-
-                    /** @var list<Error> $errors */
-                    $errors = [];
-                    $stream = null;
-                    $messageCounter = 0;
-                    $lastIndex = null;
-
-                    try {
-                        $stream = $this->messageLoader->load($startIndex, $subscriptions->toArray());
-
-                        foreach ($stream as $index => $message) {
-                            $messageCounter++;
-                            $lastIndex = $index;
-
-                            foreach ($subscriptions as $subscription) {
-                                if ($subscription->position() >= $index) {
-                                    $this->logger?->debug(
-                                        sprintf(
-                                            'Subscription Engine: Subscription "%s" is farther than the current position (%d > %d), continue processing.',
-                                            $subscription->id(),
-                                            $subscription->position(),
-                                            $index,
-                                        ),
-                                    );
-
-                                    continue;
-                                }
-
-                                $error = $this->handleMessage($index, $message, $subscription);
-
-                                if (!$error) {
-                                    continue;
-                                }
-
-                                $errors[] = $error;
-
-                                $subscriptions->remove($subscription);
-
-                                if (count($subscriptions) === 0) {
-                                    $this->logger?->info(
-                                        'Subscription Engine: No subscriptions in booting status, finish booting.',
-                                    );
-
-                                    break 2;
-                                }
-                            }
-
-                            $this->logger?->debug(sprintf(
-                                'Subscription Engine: Current event stream position: %s',
-                                $index,
-                            ));
-
-                            if ($limit !== null && $messageCounter >= $limit) {
-                                $this->logger?->info(
-                                    sprintf(
-                                        'Subscription Engine: Message limit (%d) reached, finish processing.',
-                                        $limit,
-                                    ),
-                                );
-
-                                return new ProcessedResult($messageCounter, false, $errors);
-                            }
-                        }
-                    } finally {
-                        $stream?->close();
-
-                        if ($lastIndex !== null && $messageCounter > 0) {
-                            foreach ($subscriptions as $subscription) {
-                                $error = $this->ensureCommitBatch($subscription, $lastIndex);
-
-                                if ($error) {
-                                    $errors[] = $error;
-
-                                    $subscriptions->remove($subscription);
-                                }
-
-                                $this->subscriptionManager->update($subscription);
-                            }
-                        }
-                    }
-
-                    foreach ($subscriptions as $subscription) {
-                        if ($subscription->runMode() !== RunMode::Once) {
-                            continue;
-                        }
-
-                        $subscription->finished();
-                        $this->subscriptionManager->update($subscription);
-
-                        $this->logger?->info(sprintf(
-                            'Subscription Engine: Subscription "%s" run only once and has been set to finished.',
-                            $subscription->id(),
-                        ));
-                    }
-
-                    $this->logger?->info(
-                        sprintf(
-                            'Subscription Engine: End of stream on position "%d" has been reached, finish processing.',
-                            $lastIndex,
-                        ),
-                    );
-
-                    return new ProcessedResult($messageCounter, true, $errors);
-                },
-            );
-        } finally {
-            $this->processing = false;
-        }
-    }
-
-    public function teardown(SubscriptionEngineCriteria|null $criteria = null): Result
-    {
-        $criteria ??= new SubscriptionEngineCriteria();
-
-        $this->discoverNewSubscriptions();
-
-        $this->logger?->info('Subscription Engine: Start teardown detached subscriptions.');
-
-        return $this->subscriptionManager->findForUpdate(
-            new SubscriptionCriteria(
-                ids: $criteria->ids,
-                groups: $criteria->groups,
-                status: [Status::Detached],
-            ),
-            function (SubscriptionCollection $subscriptions): Result {
-                /** @var list<Error> $errors */
-                $errors = [];
-
-                foreach ($subscriptions as $subscription) {
-                    if ($subscription->hasCleanupTasks()) {
-                        $error = $this->cleanup($subscription);
-
-                        if ($error) {
-                            $errors[] = $error;
-                        }
-
-                        continue;
-                    }
-
-                    $subscriber = $this->subscriber($subscription->id());
-
-                    if (!$subscriber) {
-                        $this->logger?->warning(
-                            sprintf(
-                                'Subscription Engine: Subscriber for "%s" to teardown or cleanup not found, skipped.',
-                                $subscription->id(),
-                            ),
-                        );
-
-                        continue;
-                    }
-
-                    $teardownMethod = $subscriber->teardownMethod();
-
-                    if (!$teardownMethod) {
-                        $this->subscriptionManager->remove($subscription);
-
-                        $this->logger?->info(
-                            sprintf(
-                                'Subscription Engine: Subscriber "%s" for "%s" has no teardown method and was immediately removed.',
-                                $subscriber::class,
-                                $subscription->id(),
-                            ),
-                        );
-
-                        continue;
-                    }
-
-                    try {
-                        $teardownMethod();
-
-                        $this->logger?->debug(sprintf(
-                            'Subscription Engine: For Subscriber "%s" for "%s" the teardown method has been executed and is now prepared to be removed.',
-                            $subscriber::class,
-                            $subscription->id(),
-                        ));
-                    } catch (Throwable $e) {
-                        $this->logger?->error(
-                            sprintf(
-                                'Subscription Engine: Subscription "%s" for "%s" has an error in the teardown method, skipped: %s',
-                                $subscriber::class,
-                                $subscription->id(),
-                                $e->getMessage(),
-                            ),
-                        );
-
-                        $errors[] = new Error(
-                            $subscription->id(),
-                            $e->getMessage(),
-                            $e,
-                        );
-
-                        continue;
-                    }
-
-                    $this->subscriptionManager->remove($subscription);
-
-                    $this->logger?->info(
-                        sprintf(
-                            'Subscription Engine: Subscription "%s" removed.',
-                            $subscription->id(),
-                        ),
-                    );
-                }
-
-                $this->logger?->info('Subscription Engine: Finish teardown.');
-
-                return new Result($errors);
-            },
-        );
-    }
-
-    public function remove(SubscriptionEngineCriteria|null $criteria = null): Result
-    {
-        $criteria ??= new SubscriptionEngineCriteria();
-
-        $this->discoverNewSubscriptions();
-
-        return $this->subscriptionManager->findForUpdate(
-            new SubscriptionCriteria(
-                ids: $criteria->ids,
-                groups: $criteria->groups,
-            ),
-            function (SubscriptionCollection $subscriptions): Result {
-                /** @var list<Error> $errors */
-                $errors = [];
-
-                foreach ($subscriptions as $subscription) {
-                    if ($subscription->isNew()) {
-                        $this->subscriptionManager->remove($subscription);
-
-                        $this->logger?->info(
-                            sprintf(
-                                'Subscription Engine: Subscription "%s" removed.',
-                                $subscription->id(),
-                            ),
-                        );
-
-                        continue;
-                    }
-
-                    if ($subscription->hasCleanupTasks()) {
-                        $error = $this->cleanup($subscription, true);
-
-                        if ($error) {
-                            $errors[] = $error;
-                        }
-
-                        continue;
-                    }
-
-                    $subscriber = $this->subscriber($subscription->id());
-
-                    if (!$subscriber) {
-                        $this->subscriptionManager->remove($subscription);
-
-                        $this->logger?->info(
-                            sprintf(
-                                'Subscription Engine: Subscription "%s" removed without a suitable subscriber.',
-                                $subscription->id(),
-                            ),
-                        );
-
-                        continue;
-                    }
-
-                    $teardownMethod = $subscriber->teardownMethod();
-
-                    if (!$teardownMethod) {
-                        $this->subscriptionManager->remove($subscription);
-
-                        $this->logger?->info(
-                            sprintf('Subscription Engine: Subscription "%s" removed.', $subscription->id()),
-                        );
-
-                        continue;
-                    }
-
-                    try {
-                        $teardownMethod();
-                    } catch (Throwable $e) {
-                        $this->logger?->error(
-                            sprintf(
-                                'Subscription Engine: Subscriber "%s" teardown method could not be executed: %s',
-                                $subscriber::class,
-                                $e->getMessage(),
-                            ),
-                        );
-
-                        $errors[] = new Error(
-                            $subscription->id(),
-                            $e->getMessage(),
-                            $e,
-                        );
-                    }
-
-                    $this->subscriptionManager->remove($subscription);
-
-                    $this->logger?->info(
-                        sprintf('Subscription Engine: Subscription "%s" removed.', $subscription->id()),
-                    );
-                }
-
-                return new Result($errors);
-            },
-        );
-    }
-
-    public function reactivate(SubscriptionEngineCriteria|null $criteria = null): Result
-    {
-        $criteria ??= new SubscriptionEngineCriteria();
-
-        $this->discoverNewSubscriptions();
-
-        return $this->subscriptionManager->findForUpdate(
-            new SubscriptionCriteria(
-                ids: $criteria->ids,
-                groups: $criteria->groups,
-                status: [
-                    Status::Error,
-                    Status::Failed,
-                    Status::Detached,
-                    Status::Paused,
-                    Status::Finished,
-                ],
-            ),
-            function (SubscriptionCollection $subscriptions): Result {
-                foreach ($subscriptions as $subscription) {
-                    $subscriber = $this->subscriber($subscription->id());
-
-                    if (!$subscriber) {
-                        $this->logger?->debug(
-                            sprintf(
-                                'Subscription Engine: Subscriber for "%s" not found, skipped.',
-                                $subscription->id(),
-                            ),
-                        );
-
-                        continue;
-                    }
-
-                    $error = $subscription->subscriptionError();
-
-                    if ($error) {
-                        $subscription->doRetry();
-                        $subscription->resetRetry();
-
-                        $this->subscriptionManager->update($subscription);
-
-                        $this->logger?->info(sprintf(
-                            'Subscription Engine: Subscriber "%s" for "%s" is reactivated.',
-                            $subscriber::class,
-                            $subscription->id(),
-                        ));
-
-                        continue;
-                    }
-
-                    $subscription->active();
-                    $this->subscriptionManager->update($subscription);
-
-                    $this->logger?->info(sprintf(
-                        'Subscription Engine: Subscriber "%s" for "%s" is reactivated.',
-                        $subscriber::class,
-                        $subscription->id(),
-                    ));
-                }
-
-                return new Result();
-            },
-        );
-    }
-
-    public function pause(SubscriptionEngineCriteria|null $criteria = null): Result
-    {
-        $criteria ??= new SubscriptionEngineCriteria();
-
-        $this->discoverNewSubscriptions();
-
-        return $this->subscriptionManager->findForUpdate(
-            new SubscriptionCriteria(
-                ids: $criteria->ids,
-                groups: $criteria->groups,
-                status: [
-                    Status::Active,
-                    Status::Booting,
-                    Status::Error,
-                ],
-            ),
-            function (SubscriptionCollection $subscriptions): Result {
-                /** @var Subscription $subscription */
-                foreach ($subscriptions as $subscription) {
-                    $subscription->pause();
-                    $this->subscriptionManager->update($subscription);
-
-                    $this->logger?->info(sprintf(
-                        'Subscription Engine: Subscription "%s" is paused.',
-                        $subscription->id(),
-                    ));
-                }
-
-                return new Result();
-            },
-        );
     }
 
     /** @return list<Subscription> */
@@ -792,7 +231,7 @@ final class DefaultSubscriptionEngine implements SubscriptionEngine, CanRefreshS
     {
         $criteria ??= new SubscriptionEngineCriteria();
 
-        $this->discoverNewSubscriptions();
+        $this->eventDispatcher->dispatch(new OnSubscriptions($criteria));
 
         return $this->subscriptionManager->find(
             new SubscriptionCriteria(
@@ -800,566 +239,5 @@ final class DefaultSubscriptionEngine implements SubscriptionEngine, CanRefreshS
                 groups: $criteria->groups,
             ),
         );
-    }
-
-    public function refresh(SubscriptionEngineCriteria|null $criteria = null): Result
-    {
-        $criteria ??= new SubscriptionEngineCriteria();
-
-        $this->discoverNewSubscriptions();
-
-        $subscriptions = $this->subscriptionManager->find(new SubscriptionCriteria(
-            ids: $criteria->ids,
-            groups: $criteria->groups,
-        ));
-
-        foreach ($subscriptions as $subscription) {
-            $subscriber = $this->subscriber($subscription->id());
-
-            if (!$subscriber) {
-                continue;
-            }
-
-            $changed = false;
-
-            if ($subscription->runMode() !== $subscriber->metadata()->runMode) {
-                $changed = true;
-                $oldRunMode = $subscription->runMode();
-                $subscription->changeRunMode($subscriber->metadata()->runMode);
-
-                $this->logger?->info(
-                    sprintf(
-                        'Subscription Engine: Subscription "%s" run mode changed from "%s" to "%s".',
-                        $subscription->id(),
-                        $oldRunMode->value,
-                        $subscription->runMode()->value,
-                    ),
-                );
-            }
-
-            if ($subscription->group() !== $subscriber->metadata()->group) {
-                $changed = true;
-                $oldGroup = $subscription->group();
-                $subscription->changeGroup($subscriber->metadata()->group);
-
-                $this->logger?->info(
-                    sprintf(
-                        'Subscription Engine: Subscription "%s" group changed from "%s" to "%s".',
-                        $subscription->id(),
-                        $oldGroup,
-                        $subscription->group(),
-                    ),
-                );
-            }
-
-            $cleanupTasks = $this->cleanupTasks($subscriber);
-
-            if ($subscription->cleanupTasks() !== $cleanupTasks) {
-                $changed = true;
-                $subscription->replaceCleanupTasks($cleanupTasks);
-
-                $this->logger?->info(
-                    sprintf(
-                        'Subscription Engine: Subscription "%s" cleanup tasks changed.',
-                        $subscription->id(),
-                    ),
-                );
-            }
-
-            if (!$changed) {
-                continue;
-            }
-
-            $this->subscriptionManager->update($subscription);
-        }
-
-        $this->subscriptionManager->flush();
-
-        return new Result();
-    }
-
-    private function handleMessage(int $index, Message $message, Subscription $subscription): Error|null
-    {
-        $subscriber = $this->subscriber($subscription->id());
-
-        if (!$subscriber) {
-            throw SubscriberNotFound::forSubscriptionId($subscription->id());
-        }
-
-        $subscribeMethods = $subscriber->subscribeMethods($message->event()::class);
-
-        if ($subscribeMethods === []) {
-            if (!isset($this->batching[$subscription->id()])) {
-                $subscription->changePosition($index);
-            }
-
-            $this->logger?->debug(
-                sprintf(
-                    'Subscription Engine: Subscriber "%s" for "%s" has no subscribe methods for "%s", continue.',
-                    $subscriber::class,
-                    $subscription->id(),
-                    $message->event()::class,
-                ),
-            );
-
-            return null;
-        }
-
-        $error = $this->checkAndBeginBatch($subscription);
-
-        if ($error) {
-            return $error;
-        }
-
-        try {
-            foreach ($subscribeMethods as $subscribeMethod) {
-                $subscribeMethod($message);
-            }
-        } catch (Throwable $e) {
-            $this->logger?->error(
-                sprintf(
-                    'Subscription Engine: Subscriber "%s" for "%s" could not process the event "%s": %s',
-                    $subscriber::class,
-                    $subscription->id(),
-                    $message->event()::class,
-                    $e->getMessage(),
-                ),
-            );
-
-            $this->handleError($subscription, $e, $message, $index);
-
-            return new Error(
-                $subscription->id(),
-                $e->getMessage(),
-                $e,
-            );
-        }
-
-        if ($this->shouldCommitBatch($subscription)) {
-            $this->logger?->debug(sprintf(
-                'Subscription Engine: Subscriber "%s" forces to commit batch.',
-                $subscription->id(),
-            ));
-
-            $this->ensureCommitBatch($subscription, $index);
-        }
-
-        if (!isset($this->batching[$subscription->id()])) {
-            $subscription->changePosition($index);
-        }
-
-        $subscription->resetRetry();
-
-        $this->logger?->debug(
-            sprintf(
-                'Subscription Engine: Subscriber "%s" for "%s" processed the event "%s".',
-                $subscriber::class,
-                $subscription->id(),
-                $message->event()::class,
-            ),
-        );
-
-        return null;
-    }
-
-    private function subscriber(string $subscriberId): MetadataSubscriberAccessor|null
-    {
-        return $this->subscriberRepository->get($subscriberId);
-    }
-
-    private function markDetachedSubscriptions(SubscriptionEngineCriteria $criteria): void
-    {
-        $this->subscriptionManager->findForUpdate(
-            new SubscriptionCriteria(
-                ids: $criteria->ids,
-                groups: $criteria->groups,
-                status: [Status::Active, Status::Paused, Status::Finished],
-            ),
-            function (SubscriptionCollection $subscriptions): void {
-                foreach ($subscriptions as $subscription) {
-                    $subscriber = $this->subscriber($subscription->id());
-
-                    if ($subscriber) {
-                        continue;
-                    }
-
-                    $subscription->detached();
-                    $this->subscriptionManager->update($subscription);
-
-                    $this->logger?->info(
-                        sprintf(
-                            'Subscription Engine: Subscriber for "%s" not found and has been marked as detached.',
-                            $subscription->id(),
-                        ),
-                    );
-                }
-            },
-        );
-    }
-
-    private function retrySubscriptions(SubscriptionEngineCriteria $criteria, Status $previousStatus): void
-    {
-        $this->subscriptionManager->findForUpdate(
-            new SubscriptionCriteria(
-                ids: $criteria->ids,
-                groups: $criteria->groups,
-                status: [Status::Error],
-            ),
-            function (SubscriptionCollection $subscriptions) use ($previousStatus): void {
-                /** @var Subscription $subscription */
-                foreach ($subscriptions as $subscription) {
-                    $error = $subscription->subscriptionError();
-
-                    if ($error === null) {
-                        continue;
-                    }
-
-                    if ($error->previousStatus !== $previousStatus) {
-                        continue;
-                    }
-
-                    if (!$this->retryStrategy($subscription)->shouldRetry($subscription)) {
-                        continue;
-                    }
-
-                    $subscription->doRetry();
-                    $this->subscriptionManager->update($subscription);
-
-                    $this->logger?->info(
-                        sprintf(
-                            'Subscription Engine: Retry subscription "%s" (%d) and set back to %s.',
-                            $subscription->id(),
-                            $subscription->retryAttempt(),
-                            $subscription->status()->value,
-                        ),
-                    );
-                }
-            },
-        );
-    }
-
-    private function discoverNewSubscriptions(): void
-    {
-        $subscriptions = $this->subscriptionManager->find(new SubscriptionCriteria());
-
-        $latestIndex = null;
-
-        foreach ($this->subscriberRepository->all() as $subscriber) {
-            foreach ($subscriptions as $subscription) {
-                if ($subscription->id() === $subscriber->metadata()->id) {
-                    continue 2;
-                }
-            }
-
-            $subscription = new Subscription(
-                $subscriber->metadata()->id,
-                $subscriber->metadata()->group,
-                $subscriber->metadata()->runMode,
-                cleanupTasks: $this->cleanupTasks($subscriber),
-            );
-
-            if ($subscriber->setupMethod() === null && $subscriber->metadata()->runMode === RunMode::FromNow) {
-                if ($latestIndex === null) {
-                    $latestIndex = $this->messageLoader->lastIndex();
-                }
-
-                $subscription->changePosition($latestIndex);
-                $subscription->active();
-            }
-
-            $this->subscriptionManager->add($subscription);
-
-            $this->logger?->info(
-                sprintf(
-                    'Subscription Engine: New Subscriber "%s" was found and added to the subscription store.',
-                    $subscriber->metadata()->id,
-                ),
-            );
-        }
-
-        $this->subscriptionManager->flush();
-    }
-
-    private function handleError(
-        Subscription $subscription,
-        Throwable $throwable,
-        Message|null $message = null,
-        int|null $index = null,
-    ): void {
-        if ($this->needRollback($subscription)) {
-            $this->rollback($subscription);
-        }
-
-        $retryStrategy = $this->retryStrategy($subscription);
-
-        if (!$retryStrategy instanceof ConditionalRetryStrategy || $retryStrategy->canRetry($subscription)) {
-            $subscription->error($throwable);
-            $this->subscriptionManager->update($subscription);
-
-            return;
-        }
-
-        $this->handleFailed($subscription, $throwable, $message, $index);
-    }
-
-    private function handleFailed(
-        Subscription $subscription,
-        Throwable $throwable,
-        Message|null $message = null,
-        int|null $index = null,
-    ): void {
-        if (!$message || $index === null) {
-            $subscription->failed($throwable);
-            $this->subscriptionManager->update($subscription);
-
-            return;
-        }
-
-        $subscriber = $this->subscriber($subscription->id());
-
-        if (!$subscriber) {
-            $subscription->failed($throwable);
-            $this->subscriptionManager->update($subscription);
-
-            return;
-        }
-
-        if ($subscriber->subscriber() instanceof BatchableSubscriber) {
-            $subscription->failed($throwable);
-            $this->subscriptionManager->update($subscription);
-
-            return;
-        }
-
-        $failedMethod = $subscriber->failedMethod();
-
-        if (!$failedMethod) {
-            $subscription->failed($throwable);
-            $this->subscriptionManager->update($subscription);
-
-            return;
-        }
-
-        try {
-            $failedMethod($message, $throwable);
-            $subscription->changePosition($index);
-            $subscription->resetRetry();
-
-            $this->subscriptionManager->update($subscription);
-        } catch (Throwable $e) {
-            $this->logger?->error(sprintf(
-                'Subscription Engine: Subscriber "%s" has an error in the failed method: %s',
-                $subscription->id(),
-                $e->getMessage(),
-            ));
-
-            $subscription->failed($throwable);
-            $this->subscriptionManager->update($subscription);
-        }
-    }
-
-    private function needRollback(Subscription $subscription): bool
-    {
-        return isset($this->batching[$subscription->id()]);
-    }
-
-    private function rollback(Subscription $subscription): void
-    {
-        if (!isset($this->batching[$subscription->id()])) {
-            throw new UnexpectedError('No batch to rollback.');
-        }
-
-        $subscriber = $this->batching[$subscription->id()];
-
-        unset($this->batching[$subscription->id()]);
-
-        $this->logger?->debug(sprintf(
-            'Subscription Engine: Subscriber "%s" rollback the batch.',
-            $subscription->id(),
-        ));
-
-        try {
-            $subscriber->rollbackBatch();
-        } catch (Throwable $e) {
-            $this->logger?->error(sprintf(
-                'Subscription Engine: Subscriber "%s" has an error in the rollback batch method: %s',
-                $subscription->id(),
-                $e->getMessage(),
-            ));
-        }
-    }
-
-    private function ensureCommitBatch(Subscription $subscription, int $index): Error|null
-    {
-        if (!isset($this->batching[$subscription->id()])) {
-            return null;
-        }
-
-        $subscriber = $this->batching[$subscription->id()];
-
-        unset($this->batching[$subscription->id()]);
-
-        $this->logger?->debug(sprintf(
-            'Subscription Engine: Subscriber "%s" commits the batch.',
-            $subscription->id(),
-        ));
-
-        try {
-            $subscriber->commitBatch();
-            $subscription->changePosition($index);
-        } catch (Throwable $e) {
-            $this->logger?->error(sprintf(
-                'Subscription Engine: Subscriber "%s" has an error in the commit batch method: %s',
-                $subscription->id(),
-                $e->getMessage(),
-            ));
-
-            $this->handleError($subscription, $e);
-
-            return new Error(
-                $subscription->id(),
-                $e->getMessage(),
-                $e,
-            );
-        }
-
-        return null;
-    }
-
-    private function checkAndBeginBatch(Subscription $subscription): Error|null
-    {
-        if (isset($this->batching[$subscription->id()])) {
-            return null;
-        }
-
-        $subscriber = $this->subscriber($subscription->id());
-
-        if (!$subscriber) {
-            return null;
-        }
-
-        $realSubscriber = $subscriber->subscriber();
-
-        if (!$realSubscriber instanceof BatchableSubscriber) {
-            return null;
-        }
-
-        $this->batching[$subscription->id()] = $realSubscriber;
-
-        $this->logger?->debug(sprintf(
-            'Subscription Engine: Subscriber "%s" starts a new batch.',
-            $subscription->id(),
-        ));
-
-        try {
-            $realSubscriber->beginBatch();
-        } catch (Throwable $e) {
-            $this->logger?->error(sprintf(
-                'Subscription Engine: Subscriber "%s" has an error in the begin batch method: %s',
-                $subscription->id(),
-                $e->getMessage(),
-            ));
-
-            $this->handleError($subscription, $e);
-
-            return new Error(
-                $subscription->id(),
-                $e->getMessage(),
-                $e,
-            );
-        }
-
-        return null;
-    }
-
-    private function shouldCommitBatch(Subscription $subscription): bool
-    {
-        if (!isset($this->batching[$subscription->id()])) {
-            return false;
-        }
-
-        return $this->batching[$subscription->id()]->forceCommit();
-    }
-
-    private function retryStrategy(Subscription $subscription): RetryStrategy
-    {
-        $subscriber = $this->subscriber($subscription->id());
-
-        if (!$subscriber instanceof MetadataSubscriberAccessor) {
-            return $this->retryStrategyRepository->getDefaultRetryStrategy();
-        }
-
-        $retryStrategy = $subscriber->metadata()->retryStrategy;
-
-        if ($retryStrategy === null) {
-            return $this->retryStrategyRepository->getDefaultRetryStrategy();
-        }
-
-        return $this->retryStrategyRepository->get($retryStrategy);
-    }
-
-    private function cleanup(Subscription $subscription, bool $force = false): Error|null
-    {
-        if (!$this->cleaner) {
-            throw new CleanerNotConfigured();
-        }
-
-        try {
-            $this->cleaner->cleanup($subscription);
-            $this->logger?->debug(
-                sprintf(
-                    'Subscription Engine: For Subscription "%s" the cleanup tasks have been executed.',
-                    $subscription->id(),
-                ),
-            );
-        } catch (Throwable $e) {
-            $this->logger?->error(
-                sprintf(
-                    'Subscription Engine: Subscription "%s" has an error in the cleanup tasks: %s',
-                    $subscription->id(),
-                    $e->getMessage(),
-                ),
-            );
-
-            if ($force) {
-                $this->subscriptionManager->remove($subscription);
-
-                $this->logger?->info(sprintf(
-                    'Subscription Engine: Subscription "%s" removed.',
-                    $subscription->id(),
-                ));
-            }
-
-            return new Error(
-                $subscription->id(),
-                $e->getMessage(),
-                $e,
-            );
-        }
-
-        $this->subscriptionManager->remove($subscription);
-
-        $this->logger?->info(sprintf(
-            'Subscription Engine: Subscription "%s" removed.',
-            $subscription->id(),
-        ));
-
-        return null;
-    }
-
-    /** @return list<object>|null */
-    private function cleanupTasks(MetadataSubscriberAccessor $subscriber): array|null
-    {
-        $method = $subscriber->cleanupMethod();
-
-        if (!$method) {
-            return null;
-        }
-
-        if (!$this->cleaner) {
-            throw new CleanerNotConfigured();
-        }
-
-        return array_values([...$method()]);
     }
 }
