+++ tests/Unit/Subscription/Engine/ThrowOnErrorSubscriptionEngineTest.php
@@ -4,12 +4,10 @@ declare(strict_types=1);
 
 namespace Patchlevel\EventSourcing\Tests\Unit\Subscription\Engine;
 
-use LogicException;
-use Patchlevel\EventSourcing\Subscription\Engine\CanRefreshSubscriptions;
+use Patchlevel\EventSourcing\Subscription\Engine\Command\Setup;
 use Patchlevel\EventSourcing\Subscription\Engine\Error;
 use Patchlevel\EventSourcing\Subscription\Engine\ErrorDetected;
 use Patchlevel\EventSourcing\Subscription\Engine\ProcessedResult;
-use Patchlevel\EventSourcing\Subscription\Engine\Result;
 use Patchlevel\EventSourcing\Subscription\Engine\SubscriptionEngine;
 use Patchlevel\EventSourcing\Subscription\Engine\SubscriptionEngineCriteria;
 use Patchlevel\EventSourcing\Subscription\Engine\ThrowOnErrorSubscriptionEngine;
@@ -20,83 +18,18 @@ use RuntimeException;
 #[CoversClass(ThrowOnErrorSubscriptionEngine::class)]
 final class ThrowOnErrorSubscriptionEngineTest extends TestCase
 {
-    public function testSetupSuccess(): void
-    {
-        $parent = $this->createMock(SubscriptionEngine::class);
-
-        $engine = new ThrowOnErrorSubscriptionEngine($parent);
-        $criteria = new SubscriptionEngineCriteria();
-
-        $expectedResult = new Result();
-
-        $parent->expects($this->once())->method('setup')->with($criteria, true)->willReturn($expectedResult);
-        $result = $engine->setup($criteria, true);
-
-        self::assertSame($expectedResult, $result);
-    }
-
-    public function testSetupError(): void
-    {
-        $this->expectException(ErrorDetected::class);
-
-        $parent = $this->createMock(SubscriptionEngine::class);
-
-        $engine = new ThrowOnErrorSubscriptionEngine($parent);
-        $criteria = new SubscriptionEngineCriteria();
-
-        $expectedResult = new Result([
-            new Error('id1', 'error1', new RuntimeException('error1')),
-            new Error('id2', 'error2', new RuntimeException('error2')),
-        ]);
-
-        $parent->expects($this->once())->method('setup')->with($criteria, false)->willReturn($expectedResult);
-        $engine->setup($criteria);
-    }
-
-    public function testBootSuccess(): void
-    {
-        $parent = $this->createMock(SubscriptionEngine::class);
-
-        $engine = new ThrowOnErrorSubscriptionEngine($parent);
-        $criteria = new SubscriptionEngineCriteria();
-
-        $expectedResult = new ProcessedResult(5);
-
-        $parent->expects($this->once())->method('boot')->with($criteria, 10)->willReturn($expectedResult);
-        $result = $engine->boot($criteria, 10);
-
-        self::assertSame($expectedResult, $result);
-    }
-
-    public function testBootError(): void
-    {
-        $this->expectException(ErrorDetected::class);
-
-        $parent = $this->createMock(SubscriptionEngine::class);
-
-        $engine = new ThrowOnErrorSubscriptionEngine($parent);
-        $criteria = new SubscriptionEngineCriteria();
-
-        $expectedResult = new ProcessedResult(5, false, [
-            new Error('id1', 'error1', new RuntimeException('error1')),
-            new Error('id2', 'error2', new RuntimeException('error2')),
-        ]);
-
-        $parent->expects($this->once())->method('boot')->with($criteria, 10)->willReturn($expectedResult);
-        $engine->boot($criteria, 10);
-    }
-
     public function testRunSuccess(): void
     {
         $parent = $this->createMock(SubscriptionEngine::class);
 
         $engine = new ThrowOnErrorSubscriptionEngine($parent);
-        $criteria = new SubscriptionEngineCriteria();
 
         $expectedResult = new ProcessedResult(5);
 
-        $parent->expects($this->once())->method('run')->with($criteria, 10)->willReturn($expectedResult);
-        $result = $engine->run($criteria, 10);
+        $command = new Setup();
+
+        $parent->expects($this->once())->method('run')->with($command)->willReturn($expectedResult);
+        $result = $engine->run($command);
 
         self::assertSame($expectedResult, $result);
     }
@@ -108,147 +41,16 @@ final class ThrowOnErrorSubscriptionEngineTest extends TestCase
         $parent = $this->createMock(SubscriptionEngine::class);
 
         $engine = new ThrowOnErrorSubscriptionEngine($parent);
-        $criteria = new SubscriptionEngineCriteria();
+
+        $command = new Setup();
 
         $expectedResult = new ProcessedResult(5, false, [
             new Error('id1', 'error1', new RuntimeException('error1')),
             new Error('id2', 'error2', new RuntimeException('error2')),
         ]);
 
-        $parent->expects($this->once())->method('run')->with($criteria, 10)->willReturn($expectedResult);
-        $engine->run($criteria, 10);
-    }
-
-    public function testTeardownSuccess(): void
-    {
-        $parent = $this->createMock(SubscriptionEngine::class);
-
-        $engine = new ThrowOnErrorSubscriptionEngine($parent);
-        $criteria = new SubscriptionEngineCriteria();
-
-        $expectedResult = new Result();
-
-        $parent->expects($this->once())->method('teardown')->with($criteria)->willReturn($expectedResult);
-        $result = $engine->teardown($criteria);
-
-        self::assertSame($expectedResult, $result);
-    }
-
-    public function testTeardownError(): void
-    {
-        $this->expectException(ErrorDetected::class);
-
-        $parent = $this->createMock(SubscriptionEngine::class);
-
-        $engine = new ThrowOnErrorSubscriptionEngine($parent);
-        $criteria = new SubscriptionEngineCriteria();
-
-        $expectedResult = new Result([
-            new Error('id1', 'error1', new RuntimeException('error1')),
-            new Error('id2', 'error2', new RuntimeException('error2')),
-        ]);
-
-        $parent->expects($this->once())->method('teardown')->with($criteria)->willReturn($expectedResult);
-        $engine->teardown($criteria);
-    }
-
-    public function testRemoveSuccess(): void
-    {
-        $parent = $this->createMock(SubscriptionEngine::class);
-
-        $engine = new ThrowOnErrorSubscriptionEngine($parent);
-        $criteria = new SubscriptionEngineCriteria();
-
-        $expectedResult = new Result();
-
-        $parent->expects($this->once())->method('remove')->with($criteria)->willReturn($expectedResult);
-        $result = $engine->remove($criteria);
-
-        self::assertSame($expectedResult, $result);
-    }
-
-    public function testRemoveError(): void
-    {
-        $this->expectException(ErrorDetected::class);
-
-        $parent = $this->createMock(SubscriptionEngine::class);
-
-        $engine = new ThrowOnErrorSubscriptionEngine($parent);
-        $criteria = new SubscriptionEngineCriteria();
-
-        $expectedResult = new Result([
-            new Error('id1', 'error1', new RuntimeException('error1')),
-            new Error('id2', 'error2', new RuntimeException('error2')),
-        ]);
-
-        $parent->expects($this->once())->method('remove')->with($criteria)->willReturn($expectedResult);
-        $engine->remove($criteria);
-    }
-
-    public function testReactivateSuccess(): void
-    {
-        $parent = $this->createMock(SubscriptionEngine::class);
-
-        $engine = new ThrowOnErrorSubscriptionEngine($parent);
-        $criteria = new SubscriptionEngineCriteria();
-
-        $expectedResult = new Result();
-
-        $parent->expects($this->once())->method('reactivate')->with($criteria)->willReturn($expectedResult);
-        $result = $engine->reactivate($criteria);
-
-        self::assertSame($expectedResult, $result);
-    }
-
-    public function testReactivateError(): void
-    {
-        $this->expectException(ErrorDetected::class);
-
-        $parent = $this->createMock(SubscriptionEngine::class);
-
-        $engine = new ThrowOnErrorSubscriptionEngine($parent);
-        $criteria = new SubscriptionEngineCriteria();
-
-        $expectedResult = new Result([
-            new Error('id1', 'error1', new RuntimeException('error1')),
-            new Error('id2', 'error2', new RuntimeException('error2')),
-        ]);
-
-        $parent->expects($this->once())->method('reactivate')->with($criteria)->willReturn($expectedResult);
-        $engine->reactivate($criteria);
-    }
-
-    public function testPauseSuccess(): void
-    {
-        $parent = $this->createMock(SubscriptionEngine::class);
-
-        $engine = new ThrowOnErrorSubscriptionEngine($parent);
-        $criteria = new SubscriptionEngineCriteria();
-
-        $expectedResult = new Result();
-
-        $parent->expects($this->once())->method('pause')->with($criteria)->willReturn($expectedResult);
-        $result = $engine->pause($criteria);
-
-        self::assertSame($expectedResult, $result);
-    }
-
-    public function testPauseError(): void
-    {
-        $this->expectException(ErrorDetected::class);
-
-        $parent = $this->createMock(SubscriptionEngine::class);
-
-        $engine = new ThrowOnErrorSubscriptionEngine($parent);
-        $criteria = new SubscriptionEngineCriteria();
-
-        $expectedResult = new Result([
-            new Error('id1', 'error1', new RuntimeException('error1')),
-            new Error('id2', 'error2', new RuntimeException('error2')),
-        ]);
-
-        $parent->expects($this->once())->method('pause')->with($criteria)->willReturn($expectedResult);
-        $engine->pause($criteria);
+        $parent->expects($this->once())->method('run')->with($command)->willReturn($expectedResult);
+        $engine->run($command);
     }
 
     public function testSubscriptions(): void
@@ -262,53 +64,5 @@ final class ThrowOnErrorSubscriptionEngineTest extends TestCase
         $result = $engine->subscriptions($criteria);
 
         self::assertSame([], $result);
-    }
-
-    public function testRefreshSubscriptionsSuccess(): void
-    {
-        $parent = $this->createMockForIntersectionOfInterfaces([
-            SubscriptionEngine::class,
-            CanRefreshSubscriptions::class,
-        ]);
-
-        $engine = new ThrowOnErrorSubscriptionEngine($parent);
-        $criteria = new SubscriptionEngineCriteria();
-
-        $expectedResult = new Result();
-
-        $parent->expects($this->once())->method('refresh')->with($criteria)->willReturn($expectedResult);
-        $result = $engine->refresh($criteria);
-
-        self::assertSame($expectedResult, $result);
-    }
-
-    public function testRefreshSubscriptionsError(): void
-    {
-        $this->expectException(ErrorDetected::class);
-
-        $parent = $this->createMockForIntersectionOfInterfaces([
-            SubscriptionEngine::class,
-            CanRefreshSubscriptions::class,
-        ]);
-
-        $engine = new ThrowOnErrorSubscriptionEngine($parent);
-        $criteria = new SubscriptionEngineCriteria();
-
-        $expectedResult = new Result([
-            new Error('id1', 'error1', new RuntimeException('error1')),
-        ]);
-
-        $parent->expects($this->once())->method('refresh')->with($criteria)->willReturn($expectedResult);
-        $engine->refresh($criteria);
-    }
-
-    public function testRefreshSubscriptionsNotSupported(): void
-    {
-        $parent = $this->createMock(SubscriptionEngine::class);
-
-        $engine = new ThrowOnErrorSubscriptionEngine($parent);
-
-        $this->expectException(LogicException::class);
-        $engine->refresh();
     }
 }
